background_tasks/parse_html_to_csv/requirements.txt
beautifulsoup4
google-cloud-storage 